Add support to '%p' format specifier, and now 'Paco' pass (not yet 'Paco -s'), adding 'write_pointer' to 'write_hexa.c' file and changing several data types of return types and parameters. Also move define constants LOWERCASE and UPPERCASE from 'ft_printf.c' to header file

+++ write_number.c
@@ -6,16 +6,16 @@
 /*   By: jfidalgo <jfidalgo@student.42bar(...).com  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/25 21:01:16 by jfidalgo          #+#    #+#             */
-/*   Updated: 2024/02/28 12:19:03 by jfidalgo         ###   ########.fr       */
+/*   Updated: 2024/02/28 15:22:38 by jfidalgo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-static long	write_number_aux(unsigned long number)
+static int	write_number_aux(unsigned long number)
 {
-	long	result1;
-	long	result2;
+	int	result1;
+	int	result2;
 
 	if (number < 10)
 		return (write_char(number + '0'));
@@ -31,9 +31,9 @@ static long	write_number_aux(unsigned long number)
 	}
 }
 
-long	write_number(long number)
+int	write_number(long number)
 {
-	long	result;
+	int	result;
 
 	if (number < 0)
 	{
